Final RAD ERP updates

+++ LiquorWarehouseStage_1_0/SFIn/Tables/gvp__RAD__c.sql
@@ -17,7 +17,7 @@
 	gvp__Distributor__c char(18) null,
 	gvp__Item__c char(18) null,
 	gvp__Label__c nvarchar(100) null,
-	gvp__Physical_Cases__c FLOAT null,
+	gvp__Physical_Cases__c DECIMAL(16, 2) null,
 	gvp__BDN_Created_Date__c date null,
 	gvp__BDN_Date_Month__c nvarchar(20) null,
 	gvp__BDN_Date_Year__c varchar(4) null,
